move classes to default maven folder, set a new package
