+++ pom.xml
@@ -4,12 +4,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>digidoggy</groupId>
-    <artifactId>scrapdun-game</artifactId>
+    <artifactId>scrapdun</artifactId>
     <version>0.1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <properties>
-        <java.version>19</java.version>
+        <java.version>11</java.version>
 
         <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
         <maven-dependency-plugin.version>3.4.0</maven-dependency-plugin.version>
