+++ src/main/java/digidoggy/scrapdun/Data.java
@@ -1,4 +1,4 @@
-package game;
+package digidoggy.scrapdun;
 
 import java.sql.*;
