+++ src/main/java/digidoggy/scrapdun/Npc.java
@@ -1,4 +1,4 @@
-package game;
+package digidoggy.scrapdun;
 
 import java.util.Scanner;
