+++ src/main/java/digidoggy/scrapdun/Player.java
@@ -1,4 +1,4 @@
-package game;
+package digidoggy.scrapdun;
 
 import java.util.ArrayList;
 import java.util.Random;
